+++ web/src/store/index.js
@@ -1,0 +1,9 @@
+import { configureStore } from '@reduxjs/toolkit'
+
+import filesReducer from './filesSlice.js'
+
+export const store = configureStore({
+  reducer: {
+    files: filesReducer
+  }
+})